Update 'Emulator.cpp' 'runProgram' function to cast op_code into int with long long while dividing

+++ src/Emulator.cpp
@@ -8,6 +8,6 @@ bool Emulator::runProgram( )
     while(true)
     {            
         long long inst = m_memory[loc];
-        int op_code = inst / 10'000'000'000;
+        int op_code = (int) (inst / 10'000'000'000LL);
     }
 }
